se-04-team-28-main/requirements.txt
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
getmac==0.8.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.1.0
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.0.1
pybase64==1.2.1
PyMySQL==1.0.2
python-arptable==0.0.2
mysql-connector-python==8.0.28
PyYAML==5.4.1
qrcode==7.3.1
Werkzeug==2.0.3
Flask-Selfdoc==1.2.0